+++ connectors/tests/config.yml
@@ -20,7 +20,7 @@ service:
 connector_id: '1'
 
 sources:
-  fake: test_sync:FakeSource
-  large_fake: test_sync:LargeFakeSource
-  fail_once: test_sync:FailsThenWork
-  fake_ts: test_sync:FakeSourceTS
+  fake: fake_sources:FakeSource
+  large_fake: fake_sources:LargeFakeSource
+  fail_once: fake_sources:FailsThenWork
+  fake_ts: fake_sources:FakeSourceTS
