+++ connectors/tests/config_https.yml
@@ -18,5 +18,5 @@ service:
 connector_id: '1'
 
 sources:
-  fake: test_sync:FakeSource
-  large_fake: test_sync:LargeFakeSource
+  fake: fake_sources:FakeSource
+  large_fake: fake_sources:LargeFakeSource
