+++ connectors/tests/config_mem.yml
@@ -19,6 +19,6 @@ service:
 connector_id: '1'
 
 sources:
-  fake: test_sync:FakeSource
-  large_fake: test_sync:LargeFakeSource
-  fail_once: test_sync:FailsThenWork
+  fake: fake_sources:FakeSource
+  large_fake: fake_sources:LargeFakeSource
+  fail_once: fake_sources:FailsThenWork
